Extract a standalone DBpedia client library
This sets up the DBClient as a standalone library that can be co-equal with a future Wikipedia library, instead of mushing it all together into tsomi-rdf.

I also do some work to improve the architecture of the Sparql web calls, including providing a promise-based API, switching to fetch, and making the code a little more manageable.

* Reformulate Sparql in terms of fetch and expose a promise-based query

  This both allows for promise-based workflows and removes jquery from the
  test suite. For me, the test suite won't run if jquery is present since
  jquery requries an actual browser.

* Add a module of HTTP utils that I commonly use

* Add eslint rules