Invert data control to D3
This follows the advice of other D3 programmers, but goes against my
architectural instincts. And, as it turns out, this leaves us with no
data appearing in the display, even though all of the DOM elements
indicate that the data should be there.

Checkpointing this on the off chance that I can show it to someone else
and they can get it working. Otherwise, I'm likely to revert to my
architecture.